+++ tests/test.c
@@ -432,6 +432,7 @@ struct Test tests[] = {
     TEST_CASE(test_minor_gc_messages),
     TEST_CASE_EXPECTED(test_raise, 7),
     TEST_CASE(test_map),
+    TEST_CASE(test_jit_map_size),
     TEST_CASE_ATOMVM_ONLY(test_refc_binaries, 0),
     TEST_CASE(test_sub_binaries),
     TEST_CASE_ATOMVM_ONLY(test_throw_call_ext_last, 0),